Fix command.js

+++ src/commands.js
@@ -254,7 +254,7 @@ function test() {
         changedTest = resume.getChangedTest();
         originalTest = resume.getOriginalTest();
         for (const singleTest of originalTest) {
-          if (!changedTest.includes(singleTest.path)) {
+          if (!changedTest.includes(singleTest.path)&&!(path.dirname(singleTest.path)===testDir+'/utils'&&!(path.dirname(singleTest.path)===testDir+'/json'))) {
             fs.unlinkSync(singleTest.path);
           }
         }
@@ -264,6 +264,7 @@ function test() {
       spawnCompile();
       var originalBytecodeMap = new Map();
       var check = false;
+      var contractsToMutate=[];
       for (const file of changedContracts) {
         if (file !== config.ignore)
           for (const contract of config.ignore) {
@@ -276,20 +277,22 @@ function test() {
           }
 
         if (check) {
-          console.log("Dir")
-          console.log(parse(file).dir)
           exploreDirectories(config.compiledDir)
           compiledContracts.map(singleContract=>{
             if(parse(singleContract).name===parse(file).name){
             originalBytecodeMap.set(parse(file).name, saveBytecodeSync(singleContract))
+            if(!contractsToMutate.includes(file)){
+               contractsToMutate.push(file)
+            }
             }})
         }
+        check=false
       }
         console.log("Contracts to mutate and Test: " + originalBytecodeMap.size);
         instrumenter.instrumentConfig();
         reporter.setupMutationsReport();
         //Generate mutations
-        const mutations = generateAllMutations(files);
+        const mutations = generateAllMutations(contractsToMutate);
         var startTime = Date.now();
 
         //Compile and test each mutant
@@ -660,13 +663,6 @@ function tce(mutation, map, file, originalBytecodeMap) {
     console.log(chalk.magenta("EQUIVALENT"));
   } else if (map.size !== 0) {
     for (const key of map.keys()) {
-      console.log('\n')
-      console.log('Map key ')
-      console.log(map.get(key))
-
-      console.log('\n\n\nmutation.bytecode')
-      console.log(mutation.bytecode)
-      console.log('\n')
       if (map.get(key) === mutation.bytecode) {
         mutation.status = "redundant";
         console.log(chalk.magenta("REDUNDANT"));
